fix(TDI-39026): should move some tests to integration test module for jdbc(#836)
