+++ components/components-jdbc/components-jdbc-integration/src/test/resources/connection.properties
@@ -1,0 +1,10 @@
+driverClass=org.apache.derby.jdbc.EmbeddedDriver
+jdbcUrl=jdbc:derby:memory:myDB;create=true
+userId=any
+password=any
+
+#driverPath=mvn:org.talend.libraries/mysql-connector-java-5.1.30-bin/6.0.0
+#driverClass=org.gjt.mm.mysql.Driver
+#jdbcUrl=jdbc:mysql://localhost:3306/test
+#userId=root
+#password=talend
